[Refactor] 店舗で現在地を偽装せず、対象の町を明示的に渡す

+++ src/store/black-market.h
@@ -1,4 +1,6 @@
 #pragma once
 
+#include <cstddef>
+
 class ItemEntity;
-bool black_market_crap(int town_num, const ItemEntity &item);
+bool black_market_crap(size_t town_num, const ItemEntity &item);
